[SOIN] Extrais les méthodes de calcul de l'ouverture et du scroll
...afin d'alléger le composant
