+++ svelte/lib/visiteGuideeSPA/VisiteGuideeSPA.svelte
@@ -13,6 +13,12 @@
   import ModaleExplicative from './ModaleExplicative.svelte';
   import { donneesEtapesVisiteGuidee } from './visiteGuideeSPA.donnees';
   import { onMount, tick } from 'svelte';
+  import {
+    ajusteHauteurScroll,
+    calculeRectangleOuverture,
+    degeleDefilementDuCorps,
+    geleDefilementDuCorps,
+  } from './visiteGuidee.utils';
 
   let { referentiel, featureFlags, nonce }: VisiteGuideeSPAProps = $props();
 
@@ -35,155 +41,31 @@
     document.getElementById('visite-guidee-cadre-blanc')!
   );
 
-  const decoupeParDefaut = { marge: 24, rayon: 8 };
-
-  let scrollGele = 0;
-
-  const geleDefilementDuCorps = () => {
-    scrollGele = window.scrollY;
-    document.body.style.overflow = 'hidden';
-    document.body.style.position = 'fixed';
-    document.body.style.top = `-${scrollGele}px`;
-    document.body.style.width = '100%';
-  };
-
-  const degeleDefilementDuCorps = () => {
-    document.body.style.position = '';
-    document.body.style.top = '';
-    document.body.style.width = '';
-    document.body.style.overflow = '';
-    window.scrollTo({ top: scrollGele, behavior: 'instant' });
-  };
+  let scrollGele = $state(0);
 
   $effect(() => {
-    geleDefilementDuCorps();
+    scrollGele = geleDefilementDuCorps();
     return () => {
-      degeleDefilementDuCorps();
+      degeleDefilementDuCorps(scrollGele);
     };
   });
 
-  const defilementActuel = () =>
-    document.body.style.position === 'fixed' ? scrollGele : window.scrollY;
-
-  const estEntierementVisible = (element: HTMLElement) => {
-    const { top, bottom } = element.getBoundingClientRect();
-    return top >= 0 && bottom <= window.innerHeight;
-  };
-
-  const seraitEntierementVisibleEnHautDePage = (element: HTMLElement) => {
-    const { top, height } = element.getBoundingClientRect();
-    const topDepuisHautDocument = top + defilementActuel();
-    return (
-      topDepuisHautDocument >= 0 &&
-      topDepuisHautDocument + height <= window.innerHeight
-    );
-  };
-
-  const scrolleSansBloquerLeCorps = (declencheScroll: () => void) => {
-    degeleDefilementDuCorps();
-    declencheScroll();
-    geleDefilementDuCorps();
-  };
-
-  const cheminRectangleArrondi = (
-    x: number,
-    y: number,
-    largeur: number,
-    hauteur: number,
-    rayon: number
-  ) => {
-    const r = Math.min(rayon, largeur / 2, hauteur / 2);
-    return (
-      `M${x + r} ${y} ` +
-      `H${x + largeur - r} ` +
-      `A${r} ${r} 0 0 1 ${x + largeur} ${y + r} ` +
-      `V${y + hauteur - r} ` +
-      `A${r} ${r} 0 0 1 ${x + largeur - r} ${y + hauteur} ` +
-      `H${x + r} ` +
-      `A${r} ${r} 0 0 1 ${x} ${y + hauteur - r} ` +
-      `V${y + r} ` +
-      `A${r} ${r} 0 0 1 ${x + r} ${y} ` +
-      `Z`
-    );
-  };
-
-  const calculeDecoupe = (
-    ouverture: HTMLElement,
-    { marge, rayon } = decoupeParDefaut
-  ) => {
-    const { left, top, right, bottom } = ouverture.getBoundingClientRect();
-    const cheminElement = cheminRectangleArrondi(
-      left,
-      top,
-      right - left,
-      bottom - top,
-      0
-    );
-    const cheminMarge = cheminRectangleArrondi(
-      left - marge,
-      top - marge,
-      right - left + marge * 2,
-      bottom - top + marge * 2,
-      rayon
-    );
-
-    rideau.style.clipPath = `path(evenodd, "M0 0 H${window.innerWidth} V${window.innerHeight} H0 Z ${cheminMarge}")`;
-    cadreBlanc.style.clipPath = `path(evenodd, "${cheminMarge} ${cheminElement}")`;
-  };
-
-  const rectEgaux = (a: DOMRect, b: DOMRect) =>
-    a.top === b.top &&
-    a.left === b.left &&
-    a.right === b.right &&
-    a.bottom === b.bottom;
-
-  const attendStabiliteRect = (
-    element: HTMLElement,
-    essaisRestants = 60
-  ): Promise<void> =>
-    new Promise((resolve) => {
-      const precedent = element.getBoundingClientRect();
-      requestAnimationFrame(() => {
-        const actuel = element.getBoundingClientRect();
-        if (rectEgaux(precedent, actuel) || essaisRestants <= 0) {
-          resolve();
-        } else {
-          resolve(attendStabiliteRect(element, essaisRestants - 1));
-        }
-      });
-    });
-
-  const desactiveTransition = (element: HTMLElement) => {
-    const transitionOriginale = element.style.transition;
-    element.style.transition = 'none';
-    element.getBoundingClientRect();
-    return () => {
-      element.style.transition = transitionOriginale;
-    };
-  };
+  const defilementActuel = $derived(
+    document.body.style.position === 'fixed' ? scrollGele : window.scrollY
+  );
 
   $effect(() => {
     (async () => {
       await donneesEtape.callbackAvantOuverture?.();
       await tick();
       const ouverture = donneesEtape.ouverture();
-      if (seraitEntierementVisibleEnHautDePage(ouverture)) {
-        if (defilementActuel() !== 0) {
-          scrolleSansBloquerLeCorps(() =>
-            window.scrollTo({ top: 0, behavior: 'instant' })
-          );
-        }
-      } else if (!estEntierementVisible(ouverture)) {
-        scrolleSansBloquerLeCorps(() =>
-          ouverture.scrollIntoView({ behavior: 'instant', block: 'center' })
-        );
-      }
-      const restaureTransition = desactiveTransition(ouverture);
-      await attendStabiliteRect(ouverture);
-      calculeDecoupe(ouverture, donneesEtape.decoupe ?? decoupeParDefaut);
-      const { top, left, width, height } = ouverture.getBoundingClientRect();
-      rectCible = { top, left, width, height };
-      restaureTransition();
+      ajusteHauteurScroll(ouverture, defilementActuel, scrollGele);
+      rectCible = await calculeRectangleOuverture(
+        ouverture,
+        rideau,
+        cadreBlanc,
+        donneesEtape.decoupe
+      );
     })();
   });
 </script>
